Project Organization
Moved headers and Source files
gitignore
