+++ .gitignore
@@ -1,0 +1,44 @@
+
+.vs/DX11-Base/v16/.suo
+.vs/DX11-Base/v16/Browse.VC.db
+.vs/DX11-Base/v16/ipch/AutoPCH/2ed6879c0c8620d8/D3D11WINDOW.ipch
+.vs/DX11-Base/v16/ipch/AutoPCH/48cab2c2675ebad4/INITIALIZE.ipch
+.vs/DX11-Base/v16/ipch/AutoPCH/638b4bea2cdd34c/HOOKING.ipch
+.vs/DX11-Base/v16/ipch/AutoPCH/6a68ab2a22e9f856/MENU.ipch
+.vs/DX11-Base/v16/ipch/AutoPCH/6db898025ab0f563/GAME.ipch
+.vs/DX11-Base/v16/ipch/AutoPCH/8875081cdfb4d8c6/CONSOLE.ipch
+.vs/DX11-Base/v16/ipch/AutoPCH/a4be26cb70e1337e/DLLMAIN.ipch
+.vs/DX11-Base/v16/ipch/AutoPCH/cc684b1cf133bd34/CONSOLE.ipch
+bin/DX11-Base.ilk
+bin/DX11-Base.dll
+x64/Debug/vc142.pdb
+x64/Debug/vc142.idb
+x64/Debug/trampoline.obj
+x64/Debug/Menu.obj
+x64/Debug/imgui_widgets.obj
+x64/Debug/imgui_tables.obj
+x64/Debug/imgui_impl_win32.obj
+x64/Debug/imgui_impl_dx11.obj
+x64/Debug/imgui_draw.obj
+x64/Debug/imgui_demo.obj
+x64/Debug/imgui.obj
+x64/Debug/Hooking.obj
+x64/Debug/hook.obj
+x64/Debug/hde64.obj
+x64/Debug/hde32.obj
+x64/Debug/Game.obj
+x64/Debug/DX11-Base.tlog/link.write.1.tlog
+x64/Debug/DX11-Base.tlog/link.read.1.tlog
+x64/Debug/DX11-Base.tlog/link.command.1.tlog
+x64/Debug/DX11-Base.tlog/DX11-Base.lastbuildstate
+x64/Debug/DX11-Base.tlog/CL.write.1.tlog
+x64/Debug/DX11-Base.tlog/CL.read.1.tlog
+x64/Debug/DX11-Base.tlog/CL.command.1.tlog
+x64/Debug/DX11-Base.log
+x64/Debug/DX11-Base.ilk
+x64/Debug/DX11-Base.dll.recipe
+x64/Debug/dllmain.obj
+x64/Debug/D3D11Window.obj
+x64/Debug/Console.obj
+x64/Debug/buffer.obj
+*.pdb
