--- Game.hpp
@@ -1,50 +1,0 @@
-#pragma once
-#include "helper.h"
-#include "Console.hpp"
-
-namespace DX11_Base {
-	class GameData
-	{
-	public:
-
-		explicit GameData();
-		~GameData() noexcept = default;
-		GameData(GameData const&) = delete;
-		GameData(GameData&&) = delete;
-		GameData& operator=(GameData const&) = delete;
-		GameData& operator=(GameData&&) = delete;
-	};
-
-	class GameVariables
-	{
-	public:
-
-		//	Dx & ImGui
-		int g_GamePid{};
-		HANDLE g_GameHandle{};
-		HWND g_GameWindow{};
-		int g_GameWidth;
-		int g_GameHeight;
-		ImVec2 g_WindowSize{};
-		LPCSTR g_GameTitle;
-		LPCSTR g_ClassName;
-		LPCSTR g_GamePath;
-
-		//	MENU RELATED
-		bool m_ShowMenu{};
-		bool m_ShowHud{};
-		bool m_ShowDemo{};
-
-		explicit GameVariables();
-		~GameVariables() noexcept = default;
-		GameVariables(GameVariables const&) = delete;
-		GameVariables(GameVariables&&) = delete;
-		GameVariables& operator=(GameVariables const&) = delete;
-		GameVariables& operator=(GameVariables&&) = delete;
-
-		void Init();
-	};
-
-	inline std::unique_ptr<GameData> g_GameData;
-	inline std::unique_ptr<GameVariables> g_GameVariables;
-}
