+++ dllmain.cpp
@@ -1,46 +1,7 @@
-// dllmain.cpp : Defines the entry point for the DLL application.
-#include "helper.h"
-#include "Console.hpp"
-#include "Game.hpp"
-#include "D3D11Window.hpp"
-#include "Hooking.hpp"
+#pragma once
+#include "initialize.hpp"
 
-DWORD WINAPI MainThread()
-{
-    using namespace DX11_Base;
-    g_Console = std::make_unique<Console>();
-
-#if DEBUG
-    g_Console->InitializeConsole("DX11 ImGui Base - Debug Console");
-    g_Console->printdbg("DX11_Base ImGui Hook - Initializing . . .\n\n", g_Console->color.DEFAULT);
-#endif
-
-    ///  ESTABLISH GAME DATA   
-    g_GameData = std::make_unique<GameData>();
-    g_GameVariables = std::make_unique<GameVariables>();
-
-    ///  CREATE WINDOW AND ESTABLISH HOOKS
-    g_D3D11Window = std::make_unique<D3D11Window>();
-    g_Hooking = std::make_unique<Hooking>();
-    g_Menu = std::make_unique<Menu>();
-    g_Hooking->Hook();
-
-#if DEBUG
-    g_Console->printdbg("Main::Initialized\n", g_Console->color.green);
-#endif
-    
-    ///  RENDER LOOP
-    g_Running = TRUE;
-    while (g_Running)
-    {
-        if (GetAsyncKeyState(VK_INSERT) & 1) g_GameVariables->m_ShowMenu = !g_GameVariables->m_ShowMenu;
-    }
-
-    ///  EXIT
-    FreeLibraryAndExitThread(g_hModule, EXIT_SUCCESS);
-    return EXIT_SUCCESS;
-}
-
+extern DWORD WINAPI MainThread_Initialize();
 BOOL APIENTRY DllMain(HMODULE hModule, DWORD  dwCallReason, LPVOID lpReserved)
 {
     using namespace DX11_Base;
@@ -49,7 +10,7 @@ BOOL APIENTRY DllMain(HMODULE hModule, DWORD  dwCallReason, LPVOID lpReserved)
     switch (dwCallReason) {
         case (DLL_PROCESS_ATTACH): {
             DisableThreadLibraryCalls(hModule);
-            CreateThread(NULL, NULL, (LPTHREAD_START_ROUTINE)MainThread, g_hModule, NULL, NULL);
+            CreateThread(NULL, NULL, (LPTHREAD_START_ROUTINE)MainThread_Initialize, g_hModule, NULL, NULL);
             break;
         }
         case (DLL_PROCESS_DETACH): {
