+++ include/Hooking.hpp
@@ -1,0 +1,23 @@
+#pragma once
+#include "helper.h"
+#include "Console.hpp"
+#include "Game.hpp"
+#include "D3D11Window.hpp"
+
+namespace DX11_Base {
+	class Hooking
+	{
+	public:
+
+		explicit Hooking();
+		~Hooking() noexcept;
+		Hooking(Hooking const&) = delete;
+		Hooking(Hooking&&) = delete;
+		Hooking& operator=(Hooking const&) = delete;
+		Hooking& operator=(Hooking&&) = delete;
+
+		void Hook();
+		void Unhook();
+	};
+	inline std::unique_ptr<Hooking> g_Hooking;
+}
